backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
google-genai==0.3.0
pydantic==2.6.4
python-dotenv==1.0.1
pytest==8.1.1
